laurel_sprout: re-import thermal config (normal)
So what xiaomi does is, use dynamic thermal profiles. That is, changing the thermal configs depending on the situation the device is in.
For eg, if any kind of heavy task is going on in the device, xiaomi switches the thermal config to no-limits (or while playing a game).

For laurel_sprout, they did the opposite. They use only one thermal config which is the 'thermal-engine.conf'.
I do reset the thermal config selection to 'thermal-engine-normal.conf' that is mapped by '0' in the sconfig. So, importing Xiaomi's 'thermal-engine.conf' in 'thermal-engine-normal.conf'.

Performance is better, ngl. Without any heating issues.

Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I7f0fc29f6e9d6cc5312557a9d9f2363de0cb18f9